tests/wpt/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-defaultTests.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  //Policy name test
  test(t => {
    assert_throws("InvalidStateError", _ => {
      let policy = window.TrustedTypes.createPolicy('default', { createHTML: s => s } );
    });
  }, "default policy has to be exposed");
</script>
